[Refactor] 라벨 뷰 분리
- 단순 라벨들을 묶어 뷰로 분리
